Change economy commands
