+++ core/src/main/java/me/lokka30/treasury/plugin/core/command/subcommand/economy/EconomyHelpSub.java
@@ -1,0 +1,15 @@
+package me.lokka30.treasury.plugin.core.command.subcommand.economy;
+
+import me.lokka30.treasury.plugin.core.command.CommandSource;
+import me.lokka30.treasury.plugin.core.command.Subcommand;
+import org.jetbrains.annotations.NotNull;
+
+// "/treasury economy help"
+public class EconomyHelpSub implements Subcommand {
+
+    @Override
+    public void execute(@NotNull final CommandSource source, @NotNull final String label, final @NotNull String[] args) {
+
+    }
+
+}
