+++ core/src/main/java/me/lokka30/treasury/plugin/core/command/subcommand/economy/migrate/MigrationEconomy.java
@@ -2,7 +2,7 @@
  * This file is/was part of Treasury. To read more information about Treasury such as its licensing, see <https://github.com/lokka30/Treasury>.
  */
 
-package me.lokka30.treasury.plugin.core.command.subcommand.migrate;
+package me.lokka30.treasury.plugin.core.command.subcommand.economy.migrate;
 
 import java.util.Collection;
 import java.util.Collections;
